# docs/api.yaml
openapi: 3.0.1
info:
  title: Lider-Cloudappi
  version: 1.0.0
  description: Lider-Cloudappi API
  contact:
    name: Cloudappi
    email: leonardo.herrera.co@cloudappi.net
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /cargar/equivalencias:
    get:
      tags:
        - Cargar
      summary: Permite obtener la tabla de equivalencias
      description: Permite obtener la tabla de equivalencias en la que relacionan los nombres de los insumos, telas, productos y su respectivo con los códigos de Siigo
      operationId: equivalencias
      responses:
        '200':
          description: Tabla de equivalencias creada con exito
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error400:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example: Bad Request
        fields:
          type: string
    Error401:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example: Unauthorized
        fields:
          type: string
...
    Error403:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example: Forbidden
        fields:
          type: string
    Error404:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example: Not Found
        fields:
          type: string
...
    Error500:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          example: Internal Server Error
        fields:
          type: string
...